+++ CHANGELOG.md
@@ -25,6 +25,7 @@
   - TREXIO interface: amplitude/determinant write functions accept `AbstractArray{<:Number}`
 * EOM-UCCSD/EOM-UDCSD and EOM-RCCSD/EOM-RDCSD (restricted to singlet excitations) methods have been implemented.
 * A new factorization of `kext` contractions for closed-shell CCSD/DCSD using a symmetric/antisymmetric representation is implemented (can be activated by setting `@set cc use_pm_kext=true`). This algorithm has two times less FLOPs than the standard implementation, however, it can be less efficient because of cache-unfriendly access in the construction of the intermediates (which is parallelized using `Threads.@threads` and should scale well with the number of threads). The standard implementation is still used by default, but the new one can be activated for testing and benchmarking purposes.
+* FCI and CIPHI work for non-Hermitian Hamiltonians.
 
 ### Fixed